+++ packages/providers/src/community/opencode/session.ts
@@ -10,6 +10,7 @@ import {
 } from './agent-config';
 import { errorMessage } from './errors';
 import type { OpencodeClientLike } from './runtime';
+import { normalizeTokens } from './tokens';
 
 let cachedLog: ReturnType<typeof createLogger> | undefined;
 
@@ -20,25 +21,6 @@ function getLog(): ReturnType<typeof createLogger> {
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === 'object' && value !== null;
-}
-
-function normalizeTokens(
-  info: Record<string, unknown> | undefined
-): import('../../types').TokenUsage | undefined {
-  const tokens = isRecord(info?.tokens) ? info.tokens : undefined;
-  if (!tokens) return undefined;
-
-  const input = typeof tokens.input === 'number' ? tokens.input : 0;
-  const output = typeof tokens.output === 'number' ? tokens.output : 0;
-  const reasoning = typeof tokens.reasoning === 'number' ? tokens.reasoning : 0;
-  const total = input + output + reasoning;
-
-  return {
-    input,
-    output,
-    ...(total > 0 ? { total } : {}),
-    ...(typeof info?.cost === 'number' ? { cost: info.cost } : {}),
-  };
 }
 
 export async function resolveSessionId(
