+++ packages/providers/src/community/opencode/tokens.ts
@@ -1,0 +1,22 @@
+import type { TokenUsage } from '../../types';
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === 'object' && value !== null;
+}
+
+export function normalizeTokens(info: Record<string, unknown> | undefined): TokenUsage | undefined {
+  const tokens = isRecord(info?.tokens) ? info.tokens : undefined;
+  if (!tokens) return undefined;
+
+  const input = typeof tokens.input === 'number' ? tokens.input : 0;
+  const output = typeof tokens.output === 'number' ? tokens.output : 0;
+  const reasoning = typeof tokens.reasoning === 'number' ? tokens.reasoning : 0;
+  const total = input + output + reasoning;
+
+  return {
+    input,
+    output,
+    ...(total > 0 ? { total } : {}),
+    ...(typeof info?.cost === 'number' ? { cost: info.cost } : {}),
+  };
+}
